refactor(feature/donkey/#50): 알림 페이지 컴포넌트 분리, 파일명 수정
