+++ yeoreum/src/pages/notifications/index.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import PostContainer from '../../components/board/PostContainer';
 import PostPageTitle from '../../components/board/PostPageTitle';
-import NotificationLists from '../../components/notifications/Notifications';
+import NotificationList from '../../components/notifications/NotificationList';
 
 function Notification() {
   return (
     <PostContainer>
       <PostPageTitle title="알림" />
-      <NotificationLists />
+      <NotificationList />
     </PostContainer>
   );
 }
